fix bool type in 'show backups' command
go-flags appears to only work with []bool, and silently discards
commands that have bool types. This was causing "show backups" to
disappear from the CLI, because of the "show-size" bool. This does a
quick fix by changing the type to []bool. At a glance, I can't see that
this argument even does anything, but since I am leaving shortly I
don't want to go any deeper than just a quick fix for now.